+++ src/test_models.hh
@@ -15,7 +15,7 @@ namespace Test{
     void makeUVSphere(GLuint& VBO_,
                      GLuint& IBO_,
                      GLuint& VAO_,
-                     size_t& numIndices_,
+                     unsigned int& numIndices_,
                      const unsigned int numSegments,
                      const unsigned int numRings);
